The dynamic update of the Fitts chart is completed.
Created a separated y axis for the methodical time values and the
experimental values. That fixed the problem.

Also integrated two slider to enable observing the graph while changing
the values of a and b. The update is dynamic.

Now, about to start styling the pages parameting section (sliders for a
and b) and adding a slider for the number of points to consider for the
plotting process.